proto: Introduce a reactor ControlHandler
This object only mission is to handle a control message (CtrlMsg). It
encapsulates a reference to the reactor and its lifetime is only the
time it takes to handle the message.

Main goal of this is to try to reduce reactor.rs size into logical
containers.

No code behavior change.

Signed-off-by: David Goulet <dgoulet@torproject.org>