rpclib: In tests, use an ersatz socketpair on windows.
Fixes #1831.

I'm doing this because switching from the socketpair crate to an
ersatz connect-to-localhost socketpair made the test pass.

My windows skills are weak, but it appears possible that you aren't
supposed to simultaneously have a blocking ReadFile and a blocking
WriteFile to a named pipe from separate threads?  At least, that's
consistent with our findings in !2758.

I first saw this hack in perl, when I needed it in 2003.
I am displeased that it is still apparently needed in 2025.