tor-rtmock: task: Replace ArcMutexData with Arc<Shared>
We're going to want to add a field inside the Arc but not inside the
Mutex, so make the Arc contents into a named-fields struct.

We don't need ArcMutexData any more.  Arc<Shared> will do fine.
Previously, we needed to implement .lock() on ArcMutexData but because
Arc<Shared> derefs to Shared, we can implement it on Shared just as
well.

The field in MockExecutor ought to be renamed, but that's textually
intrusive so will come in the next commit.