maint: Make matrix-check run in parallel in CI.
We should ideally do this for every-crate as well, but since that's
written in bash, it's a little more annoying to do. The options there
are either to rewrite it in python (probably good to do anyways) or to
modify the list_crates script to move the parallelization in there
(useful since it lets us reuse code, but doing that with a generator is
likely not possible). For now I'm just going to file a ticket to do that
and go back to working on stuff i actually should be doing :)

Fixes: #1625