README: Remove duplicate link to CONTRIBUTING.md
We already link to CONTRIBUTING.md near the top of the README in the "Links" section. There's no real need to add two additional links beyond that, especially if they're going to be in very similar looking sections. So we remove one and modify the other to look a little nicer. I've opted to keep the one closer to the end of the README since the top of the README is already covered by the "Links" section.