tor-proto: Decide which circuit leg to send cells on.
Previously, `CircuitCmd::Send` would indicate that a cell needs to be
sent down on the circuit leg that returned the `CircuitCmd`, but that's
no longer the case. Now, the `ConfluxSet` processes each `CircuitCmd`
returned by its inner `Circuit`s, and for each command it decides
whether it needs to be sent on the primary leg (if conflux is enabled
and the cell counts towards conflux sequence numbers), or on the leg it
was originally meant to (if conflux is disabled, or the cell doesn't
count towards sequence numbers). It also modifies the target hop of the
cell (because the multiplexable cells always need to be sent to the join
point).