dev doc: Respect Markdown lints in state-management-cli.md
Fixes all the lints caught by the markdownlint extension in the file
`doc/dev/notes/state-management-cli.md`, with the exception that MD024 is set to
have `siblings_only` enabled, since it makes sense to allow here. It would be
nice to set up a global markdownlint configuration some time, but that might be
tricky.