dirmgr: Always apply changes before advancing consensus.
In the directory code, we have functionality to advance the
consensus download state whenever possible, even if there is more we
could download in the current state.

That's fine, but when we're in this position, we need to be sure
that we're taking any action based on the current state (such as
installing notably parameters or, notably, protocol recommendations)
before we move on.