tor-rtmock: task: Redo subthread_block_on_future; poll on the subthread
Previously we put the future into the data structure, had the exeuctor
poll it there, and only woke up the subthread when the future was
ready.

Now, instead, we put a new TaskFutureInfo variant into the data
structure, and the executor context switches to the subthread every
time.  Then we poll the future on the subthread.

Moving the future into the data structure meant we had to do use a
proxy future to store the return value.  Now that's not needed any
more.

Now the future doesn't need to be Send + 'static.

Fixes #1933.