cell: Refine StreamIdReq for V1 relay cells, with minimal changes
Fortunately, the type is crate-internal, so this is not a breaking
change.

The key insight with this change is that, from the _caller's_ POV,
the set of cells that are allowed to have StreamIDs has not changed.
The only difference is that the V1 relay cell code needs to
distinguish unrecognized commands from the SENDME case.