cell, proto: Use correct Data sizes for v1 relay cells
Since v1 cells have a longer tag, they can fit less data into a
single cell.  Ah well, that's the cost of improved security.

The code in data.rs is a little wonky, in that it currently requires
its buffer to be exactly the maximum size for a data cell.  We have
a TODO about fixing that in the future, but for now I've moved it to
use a boxed slice rather than a boxed array.

Part of #1944.