proto: Avoid panic on non-existent hop number
In `send_relay_cell()` in `tunnel/reactor/circuit.rs`,
replace an unconditional array access (which would cause a panic if
`hop_num` were out-of-range) with a checked `get_mut()` call.

It's not totally clear whether this can happen in practice,
but in either case, an error is probably better than a panic.
All of our other lookups in this vector are either checked,
or more obviously infallible.

Closes #1950.