hyper-examples: Fix example hyper-http-hs-example
This example previously remained stuck in the while loop without any output, seeming as if this example is not working.

This because the method `is_fully_reachable` is no guarantee for the actual status of the onion service. It might return false while the service is actually reachable.

We added a timeout to not remain stuck in the loop.