Emit SendmeTag directly from cell crypto.
This changes the code to copy a SendmeTag rather than returning a
slice. This isn't actually a big change: sending a slice already
required 16 bytes (on 64-bit platforms), so sending a SendmeTag
around isn't a big deal.

We rely extensively on the compiler's ability to optimize away
all the checking in code like this:

```
 let slice: &[u8];
 let a: [u8;N] = slice[0..N].try_into().expect("Nope");
```

I've spot-checked it somewhat with "cargo-show-asm", but
it could use more thorough checking.

Closes #1956.