proto, cell: Remove RelayCellFormatTrait.
The purpose of the trait was to parameterize the tor1 cell crypto
on the different possible relay cell layouts.

It made sense to have this trait when we thought we would implement
the new cell layout for prop340 (packed-and-fragmented) well before
we implemented CGO.

But it now appears all but certain that CGO will land long before
we make any more headway on prop340.  Therefore,
it doesn't make sense to carry the ability to customize `tor1`
for other relay cell layouts.

Removing this trait saves a fair bit of complexity.