proto: Apply type-specific wrappers for tor1 crypto
Now instead of using CryptState for everything, we have specific
types for each role and direction of crypto.

This turned up a harmless-so-far bug in our onion service code: as
an onion service, we were using _client_ crypto layers to respond to
a client request.  That's not correct, and wouldn't have worked
with CGO.  Instead, we need to use relay crypto layers, wrapped
as client layers.

Closes #1975.