tunnel: Implement Buildable for ClientTunnel
In order to pull this off, the Arc requirement needs to go away because
the Arc<ClientCirc> is now within the ClientTunnel.

This commit also has a rename of the CircuitBuilder to TunnelBuilder in
order to reflect the change that it now builds a ClientTunnel.

There is a slight rename in tor-proto as well just for accuracy.

Signed-off-by: David Goulet <dgoulet@torproject.org>