circmgr: Make path module public on "--features=experimental-api"
Back in 4c1eb94173521bc5104449327650e20ffe32afa7, for sensible
reasons, we made `tor_circmgr::path` a crate-private module.  But
when we did that, we lost the ability for callers to construct
circuits with custom paths.

This will make it possible for callers to build custom circuits
again, without committing to a very-long-term API for that.

Closes #1981.