tor-proto: Move CircHop to a separate module.
This helps hide the `CircHop` internals, and is the first step towards
providing a safer API that aims to reduce contention and prevent
deadlocking on the stream map mutex.

This change is also in preparation for implementing special handling for
the join point of a conflux tunnel (which will involve adding a new
`CircHop` API for sharing the stream map of another `CircHop`).

I recommend reviewing this diff with `--color-moved`.