proto: Move ClientCirc stream functions to ClientTunnel
In order to pull this off, some client => tunnel renaming needed to
happen including the comments.

The send_raw_msg() is an experimental and expert mode method that any
tunnel should have access to in order to be able to send whatever
message in whatever tunnel type.

No behavior changes.

Signed-off-by: David Goulet <dgoulet@torproject.org>