tor-proto: Fix bug where the join point is read from the wrong list.
If we reach this point and the join point is `None`, it means the
conflux set has so far consisted of a single leg. This means we need to
assign the last hop of the (only) leg to the join point. This initial
leg is in `self.circuits`, not in `legs` (`legs` is the list of *new*
legs that are being added to the set).