shadow CI: have arti processes log to file instead of console
This is in preparation to switch arti's console logging to use stderr
instead of stdout. The test script currently fails the test if there are
non-empty stderr files, since that's where rust panics end up getting
recorded.

It will also make it easier to see if there is any other non-logging
output.

Also disable trace-level logging for the tor_proto module. It's unclear
exactly when/why this was added, but it's probably overkill other than
when debugging.