NetDir::pick_n_relays: avoid some unwarranted warnings
Previously we logged a warning whenever  `choose_multiple_weighted`
returns `Ok(iter)`, whenever `iter` contains fewer relays than
requested.

However, this is expected behavior of `choose_multiple_weighted` if the
filtered list of relays doesn't have the requested number of relays, in
which case it should return all of the relays in the filtered list.
In particular, this condition occurs regularly in test networks when
looking for candidate relays to add to the potential guard set while
excluding those already in the potential guard set.

We still warn if the returned list is fewer than requested, *and* fewer
than available in the filtered list.