tor-proto: Add a new identifier type for tunnels.
Currently, a tunnel is uniquely identified by the `UniqId` of the first
circuit added to the tunnel. This works, but the double-meaning of the
`UniqId` is bound to cause confusion in the future (because it blurs the
distinction between tunnels and circuits).

This introduces a new `TunnelId` type which will replace `UniqId` in the
tunnel reactor.