proto: Make conflux mock exit send SENDMEs at the right time.
We now send the SENDMEs as per the client's expectations, using the
right tags. This makes the test less fragile and easier to extend.

Note: the `multipath_stream` test currently fails, because
the mock exit endpoint never actually sends an END, so the stream never
gets closed:

```
all futures blocked. waiting for the real world? or deadlocked (waiting for each other) ?
```

This will be fixed in a future commit.