proto: Put mock runtime behind mutex.
The mutex prevents concurrent uses of
`MockRuntime::advance_until_stalled` (which are forbidden by the mock
runtime), enabling us to call this (and its functions that advance time)
more liberally in the future.

Note: the `multipath_stream` test is *still* broken, but slightly less
so. Now the only failure is "all futures blocked. waiting for the real
world? or deadlocked (waiting for each other) ?", which can be solved by
ensuring the mock exit tasks always exit: now that the mock client task
keeps the stream alive until it receives an END cell, we need to ensure
the exits tasks can complete their own, without relying on the client
to end the stream. This will be fixed in another commit.