proto: Sort the received test data.
The conflux client now sends data to the mock exit relay over both legs
(it SWITCHes legs at some point), which causes the test to fail
(because the two legs of the mock exit write the data racily to the same
`Vec`, without attempting to put it in the right order).

To work around the lack of handling of out-of-order cells at the mock
exit, we can just sort the received data to make sure we got it all.