proto: Add helper function to resolve TargetHop
The Reactor has two functions to resolve both TargetHop into a
HopLocation and then a HopLocation into a (UniqId, HopNum).

This commit simply adds a helper that does both but returns an Option
instead of a Result as it will be used by the reactor command handler
and more in future conflux commits.

Signed-off-by: David Goulet <dgoulet@torproject.org>