proto: Add a control command to get the binding key
This allows us to use TargetHop instead of HopNum but also to get one
step closer to not depend on a mutable state.

We prefer resolving a TargetHop within the Reactor object in order to
use the circuit list instead of the MutableState path.

The HS service subsystem is modified to use this modified function that
is now async and uses a TargetHop.

Signed-off-by: David Goulet <dgoulet@torproject.org>