proto: Make send_raw_msg() use a TargetHop
In order for this to work, a last_target_hop() function is added to
ClientCirc in order to return a precise hop location as a TargetHop of
the last hop.

This is needed because in the HS subsystem, we need such value in order
to get a location on the last physical hop before adding the virtual
hop.

The RDV1 cell is sent to that last target hop while the
allow_stream_request() is done on the virtual target hop.

Signed-off-by: David Goulet <dgoulet@torproject.org>