proto: Fix conflux-related congestion control proto violation.
Previously, if the primary leg was blocked on cc, the conflux set would
continue polling the stream map non-blocked leg, and send a
`CircuitCmd::Send` instructing the reactor to send a DATA cell on the
non-blocked leg. The reactor would then (wrongly) reroute the DATA cell
to the primary leg, in violation of congestion control.