proto: Ensure we get all the expected SWITCH cells.
Previously, this failed to assert that the mock relays received all the
expected SWITCH cells.

As it turns out, the assertion currently fails, because the default
value for our estimated RTTs is no longer zero (as of !3074), so we no
longer unnecessarily switch legs as much as we used to.