hsservice: Stop using HopNum and use TargetHop
This requires some changes to the tor-proto crate to handle the inbound
TargetHop from the HS subsystem and then resolve it into a HopNum for a
single circuit.

It is expected that this will change again with Conflux to only use
HopLocation internally in a Tunnel and then use HopNum into a Circuit.

Signed-off-by: David Goulet <dgoulet@torproject.org>