proto: Make RttEstimator::{ewma,min}_rtt_usec() return `Option`.
A `None` value signals to the conflux code to know to fall back on the
initial RTT of the circuit.

Without this change, the conflux switching logic is broken as we end up
staying on the leg with the best initial RTT forever (the other leg is
never picked, because its `ewma_rtt()` is stuck on `u32::MAX`, and never
updated as we never send on it).