tor-hsservice: Make num_dequeued u32 instead of usize.
We need to divide a Duration by this, and Duration only supports
division by u32.

This is, in a sense, just pushing around where the overflow would
happen, (from the conversion in the suggested update to the increment
when a item is dequeued). However, this overflow is so unlikely to
happen (it would require more that 14 million requests per second) that
it does not seem worth slowing down the increment operation to try to
handle it.