proto: Set is_multi_path = true in the tests.
This fixes the `allow_stream_requests()` tests which ensures we don't
allow incoming streams on multipath tunnels (because we don't currently
support it).

This also changes `newcirc_ext` to return `ClientTunnel` instead of
`Arc<ClientTunell>` (we need to mutate `ClientTunnel` to manually set
`is_multi_path` on its inner `ClientCirc`).