tor-proto: Temporary measure to disable CGO.
This commit prevents us from ever deciding that another relay
supports CGO, which prevents us from trying to negotiate it.
It also adds an assertion to make sure that we haven't tried to
negotiate it.

We can remove this once we're ready to have CGO turned on:
right now it's blocked on being able to negotiate CC.

I'm adding this so we can merge this branch (so that I can stop
rebasing it.)