tor-hsservice: Remove outdated TODO.
I have thought about this and come to the conclusion (which is what I
suspected when I wrote it) that the current behaviour is correct.

The attack described is completely impractical (the space of nonces is
very large), and checking whether a nonce is a replay is cheaper than
verifying a PoW solve, so we want to do that first.

Splitting this into something like the following:

* Check replay log without updating
* Check that solve is valid
* Update replay log

Would require adding a somewhat dangerous API to the ReplayLog, and
requires doing more work per request for something that isn't even a
practical attack, AFAICT.