doc/dev/notes: relay-reactor: Add a relay reactor implementation sketch.
This is a rough sketch for the new relay tunnel reactor design.
It also contains a proposed refactoring which would enable us to
repurpose much of the existing tunnel reactor implementation.

This is by no means final, and we might choose not to refactor the
tunnel reactor until after we have an initial working (or quasi-working)
version for the relay reactor.