circmgr: Don't count channel negotiation towards circuit built time.
This is important, since some circuits require us to negotiate a new
channel, and some don't: by adding this variability, we can easily
learn a timeout based on pre-existing channels to our guards, but
then later time out if it takes a while to open a TLS connection.

This is a possible solution to part of #2079, a bug where we "learn"
a circuit timeout that is too low for us to satisfy.