tor-error: allow only error given to `_report!` macros
This is useful when you create an `internal!` error and then immediately
report it. If the `_report!` macro also requires a message, then you
need to provide two messages for the error, which doesn't always make
sense.

This is already possible anyways with `warn_report!(e,)`.
Now you don't need the comma.