proto: Add a placeholder for the relay reactor.
This is a placeholder, and will likely change quite a bit in the near
future. In particular, much of this is copied from the client tunnel
reactor (a future change will refactor both of them to reduce/minimize
code duplication).

I'm adding this placeholder because the channel code will soon
need the ability to create and launch circuit reactors.