hsservice: Move derive_more::From out of internal_prelude.
This fixes an error from nightly. The trouble is that with nightly,
there's a now a [derive macro for From][issue].  That doesn't cause
a conflict when we `use derive_more::From`, but it _does_
cause a conflict when we import `derive_more::From` via
`use internal_prelude::*`.

So as a solution, we just import `derive_more::From` explicitly.

Closes #2124

[issue]: https://github.com/rust-lang/rust/pull/144922