proto: Allow padding in all channel message sets
First of all, VPADDING has been added in link protocol version 3 so it
was missing from v4.

Second, after closely looking at C-tor and the spec, it appears that we
allow VPADDING at any point on a channel which should simply be silently
dropped. Any number in any order.

Third, couple sets were missing the PADDING cell which is only allowed
on an open channel.

Signed-off-by: David Goulet <dgoulet@torproject.org>