proto: Only send the bad test cell on one of the circuit legs.
This makes it clearer that the tunnel shutdown is triggered correctly
(i.e. that it is enough for *one* of the legs to receive the bad SWITCH
cell).

This also updates a comment that wrongly suggests that the tunnel is
only shut down if both legs get removed from the conflux set.