proto: Put in a ChannelFrame<T> into the client handshake
The ClientInitiatorHandshake holds a "tls" sink but the very first thing
we do is transform it to a ChannelFrame<T>. Instead, just store the
frame to the object directly so we can then use a channel frame
uniformily accross its lifetime.

This will be useful for the future refactoring paving the way for relay
channel authentication.

Signed-off-by: David Goulet <dgoulet@torproject.org>