proto: Introduce a ChannelBaseHandshake trait
Client and relay handhsake share a lot of code because they both
send/recv the same cells, just handles them differently for
verification.

This is the base trait for all handshake implementing basic getters and
VERSIONS cell handling.

This will allow the RelayInitiatorHandshake and RelayResponderHandshake
to use this common code.

See, traits are fun. Win-win-win.

Signed-off-by: David Goulet <dgoulet@torproject.org>