proto: Use allow(unused) instead of expect(unused).
With rustc 1.85, this `expect(unused)` triggers a "lint expectation
unfulfilled" error, because rust thinks the struct *is* used (because of
the d-d `TryFrom` implementation), despite the fact that it's never
actually constructed. This isn't a problem on newer versions (it's
correctly identified as unused/dead code on 1.89, for example).

Using `allow` instead of `expect` is slightly less nice, but it makes
`clippy` pass on 1.85 too.