shadow tests: use a randomized simulation seed
Mitigates #2154

The determinism inside CI is of limited usefulness since a CI seed
generally won't perfectly replicate the behavior when debugging locally.
Arguably it's a feature that when there's some flakiness it does
replicate inside of CI, but it leads to some confusion about blame (e.g.
as in #2154 where an unrelated MR probably perturbed the PRNG to trigger
some pre-existing flakiness), and means that we can't just rerun the job
if we get really unlucky. i.e. even if we set a lower success threshold,
a sufficiently unlucky MR can still fail and keep failing as long as it
runs with the same seed.