proto: Experimental API to install a circuit padder at a hop.
For now, we're not going to implement hop negotiation: we're still
waiting to find a set of padding machines that will be useful for
Arti.  This API will help facilitate experimentation to find such
padding machines.

These APIs are under a separate feature (`circ-padding-manual`)
to avoid confusing them with the rest of the padding code:
we only want them to be exported when the researcher wants to
manually install circuit padding.