tor-hsservice: add config to disable service
Added the `enabled` field to the config options for hidden services. The
default is `true`. If it is set to `false`, the service won't start. As
of this specific commit, it's a hard error; it'll be checked again at
an earlier stage where it won't be once the feature is ready.

Signed-off-by: hashcatHitman <3924-hashcatHitman@gitlab.torproject.org>