opentelemetry: Panic if configured for non-localhost HTTP endpoint.
We may want to add a way to opt out of this check, if someone asks for
it. Checking for "localhost" and "127.0.0.1" is not a very robust check
for a loopback interface, and it's possible that there could be other
situations we can't detect where HTTP is safe (private network,
WireGuard tunnel, etc). But I don't expect those to be likely to come up
in practice, so this should be fine for now.