proto: Add test to ensure half-streams are removed after a while.
This tests that the half-stream expiry works as expected.

Note: it doesn't! This test currently fails, because there's a bug in
the way half-streams are expired. Because we don't do it on a timer, and
instead garbage-collect the half-streams on each reactor iteration, if
the reactor is stuck long enough `.await`ing a message on one of its
channels (for example, the `input` one), there is a chance it will
accept a cell on a half-stream that should've been expired.

A future commit will fix this bug.