arti: fix hs "enabled" in example config
I didn't realize we had tests that made use of the example config. The
pre-commit and pre-push hooks I had didn't trigger them, so my
modification to the example config silently got through and caused
`cfg::test::onion_services` to fail in CI.

The issue was that I had put my example for
`onion_services."allium-cepa".enabled` with a default of `true`. The
correct default is actually `"auto"`, so I switched it and improved the
description of the options.

I've fixed my hooks. Which, to be fair, are just the hooks provided in
`./maint/hooks/`. Maybe those should be improved?

Signed-off-by: hashcatHitman <3924-hashcatHitman@gitlab.torproject.org>